Add an explicit /competitions route.
